chore: refactor project into `wasp` and `wasp_backend_*` static libraries
The purpose of seperating the two is so that the linker can perform any normal optimizations on the `wasp` library (therefore dropping any symbols that the executable you're linking to does not use), but we force the linker to entirely include whichever `wasp_backend_*` you wish to use.

The entire backend library must be included since we have to register the backend as a "plugin" in a seperate data section that the linker otherwise will discard. This caused the executable to fail since it can't find any registered backend.